Finish implementing new database update to all files
Fix dashboard & all of the panels, admin, plan, member, trainer, etc.

+++ GGG_System/crud/delete-class.php
@@ -1,12 +1,19 @@
 <?php
 include('../config/constant.php');
+include('../partials/login-check.php');
+include('../functions.php');
 
 if(isset($_GET['id']))
 {
     $id = $_GET['id'];
 
-    $sql = "DELETE FROM class where class_id = $id";
+    if(checkMemberClass($conn, $id) == true){
+        $_SESSION['Foreign'] = "<br><span style = 'color: red; font-weight: bold;'> Error ... Unable to Delete Class.</span>";
+        header('location:'.SITEURL.'manage-class.php');
+        die();
+    }
 
+    $sql = "DELETE FROM class where class_id = $id";
     $res = mysqli_query($conn, $sql);
 
     if($res == true)
