+++ GGG_System/crud/delete-trainer.php
@@ -2,10 +2,17 @@
 
 include('../config/constant.php');
 include('../partials/login-check.php');
+include('../functions.php');
 
 if(isset($_GET['id']) AND isset($_GET['image'])){
     $id = $_GET['id'];
     $image = $_GET['image'];
+
+    if(checkTrainer($conn, $id) == true){
+        $_SESSION['Foreign'] = "<br><span style = 'color: red; font-weight: bold'>Error ... Unable to delete Trainer.</span>";
+        header('location:'.SITEURL.'manage-trainer.php');
+        die();
+    }
 
     $sql = "DELETE FROM user where user_id = $id";
